+++ es/util.go
@@ -70,6 +70,12 @@ func normalizeDestination(tpl map[string]interface{}) {
 	delete(tpl, "last_update_time")
 }
 
+func normalizePolicy(tpl map[string]interface{}) {
+	delete(tpl, "last_updated_time")
+	delete(tpl, "error_notification")
+	delete(tpl, "policy_id")
+}
+
 func normalizeIndexTemplate(tpl map[string]interface{}) {
 	delete(tpl, "version")
 	if settings, ok := tpl["settings"]; ok {
@@ -267,4 +273,38 @@ func (ds *resourceDataSetter) set(key string, value interface{}) {
 		return
 	}
 	ds.err = ds.d.Set(key, value)
+}
+
+func expandIndexPermissionsSet(resourcesArray []interface{}) ([]IndexPermissions, error) {
+	vperm := make([]IndexPermissions, 0, len(resourcesArray))
+	for _, item := range resourcesArray {
+		data, ok := item.(map[string]interface{})
+		if !ok {
+			return vperm, fmt.Errorf("Error asserting data as type []byte : %v", item)
+		}
+		obj := IndexPermissions{
+			IndexPatterns:  expandStringList(data["index_patterns"].(*schema.Set).List()),
+			Fls:            expandStringList(data["fls"].(*schema.Set).List()),
+			MaskedFields:   expandStringList(data["masked_fields"].(*schema.Set).List()),
+			AllowedActions: expandStringList(data["allowed_actions"].(*schema.Set).List()),
+		}
+		vperm = append(vperm, obj)
+	}
+	return vperm, nil
+}
+
+func expandTenantPermissionsSet(resourcesArray []interface{}) ([]TenantPermissions, error) {
+	vperm := make([]TenantPermissions, 0, len(resourcesArray))
+	for _, item := range resourcesArray {
+		data, ok := item.(map[string]interface{})
+		if !ok {
+			return vperm, fmt.Errorf("Error asserting data as type []byte : %v", item)
+		}
+		obj := TenantPermissions{
+			TenantPatterns: expandStringList(data["tenant_patterns"].(*schema.Set).List()),
+			AllowedActions: expandStringList(data["allowed_actions"].(*schema.Set).List()),
+		}
+		vperm = append(vperm, obj)
+	}
+	return vperm, nil
 }